DomElementWrapper is private since it used only in Base
